Comments, refactoring for variable names and initial read from file

+++ Configuration.h
@@ -8,9 +8,11 @@
 #ifndef CONFIGURATION_H_
 #define CONFIGURATION_H_
 
+#include "Arduino.h"
+
 class Configuration {
 public:
-	unsigned int mac[6] = {0x00, 0x12, 0xFB, 0x95, 0x59, 0xCF};
+	byte mac[6] = {0x00, 0x12, 0xFB, 0x95, 0x59, 0xCF};
 	unsigned int mqttServerIP[4] = {192, 168, 1, 190};
 	int mqttServerPort = 1883;
 	int AFTER_CHANGE_DELAY = 100;
